tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/events/018.html

<!doctype html>
<title>WebSockets: toString(), bubbles, cancelable</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo_raw');
  ws.addEventListener('open', t.step_func(function(e) {
    // first a text frame, then a frame with reserved opcode 3
    // which should fail the connection
    ws.send('\\x81\\x04test\\x83\\x03LOL');
    assert_equals(e.toString(), '[object Event]', "open e.toString()");
    assert_equals(e.bubbles, false, 'open e.bubbles');
    assert_equals(e.cancelable, false, 'open e.cancelable');
  }), false);
  ws.addEventListener('message', t.step_func(function(e) {
    assert_equals(e.toString(), '[object MessageEvent]', "message e.toString()");
    assert_equals(e.bubbles, false, 'message e.bubbles');
    assert_equals(e.cancelable, false, 'message e.cancelable');
  }), false);
  ws.addEventListener('error', t.step_func(function(e) {
    assert_equals(e.toString(), '[object Event]', "error e.toString()");
    assert_equals(e.bubbles, false, 'error e.bubbles');
    assert_equals(e.cancelable, false, 'error e.cancelable');
  }), false);
  ws.addEventListener('close', t.step_func(function(e) {
    assert_equals(e.toString(), '[object CloseEvent]', "close e.toString()");
    assert_equals(e.bubbles, false, 'close e.bubbles');
    assert_equals(e.cancelable, false, 'close e.cancelable');
    t.done();
  }), false);
});
</script>
